[Tip Calculator] Updated Styling & Added Icons
Added icons for the Cost, Quality of Service, and Tip Rounding sections.
Aligned items with icons and increased spacing to reduced crowdedness.
Added a style to make text and text boxes more consistent. (Note:  Style attributes can be overridden by specifying them again when the instance of the object is made)